% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{register_knit_print}
\alias{register_knit_print}
\title{Register a knit_print() method for grViz objects}
\usage{
register_knit_print()
}
\description{
This function registers a method for \code{\link[=knit_print]{knit_print()}} for \link{grViz} objects.
With this method enabled, rendered HTML documents no longer require
a viz.js dependency.
Instead, embedded SVG is inserted into the Markdown.
The resulting HTML documents have a smaller footprint and are rendered faster.
}
